+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Added debug log emits for the ``DataFrameETL`` transformer (#24, #27)
+- Emit a warning if the user attempts to expand a column with
+  too many categories (#25, #26)
 
 ## [0.1.6] - 2018-1-12